refactor user code to handler too
